restructure to clean architecture
