--- apps/prisma/go.mod
@@ -1,12 +1,0 @@
-module github.com/stamford-syntax-club/course-compose/prisma
-
-go 1.21.2
-
-require (
-	github.com/iancoleman/strcase v0.3.0
-	github.com/joho/godotenv v1.5.1
-	github.com/shopspring/decimal v1.3.1
-	github.com/steebchen/prisma-client-go v0.28.0
-	github.com/takuoki/gocase v1.1.1
-	golang.org/x/text v0.14.0
-)
