--- apps/reviews-api/.env.development
@@ -1,1 +1,0 @@
-DATABASE_URL=postgres://syntax:stamford@localhost:5432/coursecompose-dev
