+++ apps/reviews-api/Dockerfile
@@ -1,18 +1,20 @@
 FROM golang:alpine as build-base
 
-WORKDIR /review-app
+WORKDIR /review-api
 
-# Generate Prisma
-RUN mkdir -p ../prisma
-COPY ./packages/prisma/schema.prisma ../prisma/
-COPY ./packages/prisma/.env.docker ../prisma/.env
-COPY ./apps/prisma/go.mod ./apps/prisma/go.sum ../prisma/
-RUN cd ../prisma && go run github.com/steebchen/prisma-client-go generate
+# Prisma
+COPY ./packages/prisma/schema.prisma .
+COPY ./packages/prisma/.env.docker .env
 
-# Build Reviews API
 COPY ./apps/reviews-api/go.mod ./apps/reviews-api/go.sum ./
 COPY ./apps/reviews-api ./
-RUN go mod download && CGO_ENABLED=0 GOOS=linux go build -o ./out/reviews-api ./cmd
+
+# Generate Prisma Client
+RUN PRISMA_PROVIDER='go run github.com/steebchen/prisma-client-go' PRISMA_OUTPUT='./review/data/datasource/db' go run github.com/steebchen/prisma-client-go generate --schema=./schema.prisma
+ 
+
+# Build Reviews API
+RUN go mod download && CGO_ENABLED=0 GOOS=linux go build -o ./out/reviews-api main.go
 
 # Optimize Image Size
 FROM alpine:3.16.2
@@ -22,8 +24,8 @@ WORKDIR /app
 ARG ENVIRONMENT
 ENV ENVIRONMENT=$ENVIRONMENT
 
-COPY --from=build-base /review-app/out/reviews-api ./
-COPY --from=build-base /review-app/go.mod ./
+COPY --from=build-base /review-api/out/reviews-api ./
+COPY --from=build-base /review-api/go.mod ./
 # go.mod is needed for determining the root dir
 
 EXPOSE 8003
