+++ apps/reviews-api/Dockerfile.integration
@@ -1,11 +1,9 @@
 FROM golang:alpine
 
-WORKDIR /go/src/review-api
+WORKDIR /review-app-integration
 
-COPY . .
+COPY ./packages/prisma/schema.prisma .
+COPY ./apps/reviews-api .
 
-RUN go run github.com/steebchen/prisma-client-go generate 
-
-# Wait for prisma bootstrap to finish
-CMD sh -c "sleep 10 && go test -v -tags=integration ./..."
+CMD sh -c "go run github.com/steebchen/prisma-client-go generate & go test -v -tags=integration ./..."
