+++ apps/reviews-api/docker-compose.integration.yaml
@@ -11,6 +11,8 @@ services:
             POSTGRES_DB: coursecompose-integration
             POSTGRES_USER: syntax
             POSTGRES_PASSWORD: stamford
+        volumes:
+            - ./review/data/datasource/db/init.sql:/docker-entrypoint-initdb.d/init.sql
         healthcheck:
             test: ["CMD-SHELL", "pg_isready"]
             interval: 10s
@@ -19,31 +21,16 @@ services:
         networks:
             - reviews-test
 
-    prisma-bootstrap:
-        build:
-            context: ../../ 
-            dockerfile: Dockerfile.dev 
-            args:
-                DATABASE_URL: "postgres://syntax:stamford@postgres-integration:5432/coursecompose-integration"
-                PRISMA_PROVIDER: "prisma-client-js"
-                PRISMA_OUTPUT: "/app/node_modules/.prisma/client"
-        depends_on:
-            - postgres-integration
-        networks:
-            - reviews-test
-
     integration-test:
         build:
-          context: .
-          dockerfile: Dockerfile.integration
-        volumes:
-            - ../prisma:/go/src/prisma
-            - ../../packages/prisma:/go/src/prisma
+            context: ../../
+            dockerfile: apps/reviews-api/Dockerfile.integration
         environment:
             - DATABASE_URL=postgres://syntax:stamford@postgres-integration:5432/coursecompose-integration
+            - PRISMA_PROVIDER=go run github.com/steebchen/prisma-client-go
+            - PRISMA_OUTPUT=/review-app-integration/review/data/datasource/db
             - JWT_SECRET=test-jwt-token
         depends_on:
             - postgres-integration
-            - prisma-bootstrap
         networks:
             - reviews-test
