+++ apps/reviews-api/docker-compose.yaml
@@ -1,0 +1,36 @@
+version: "3.8"
+
+networks:
+    kafka:
+
+services:
+    zookeeper:
+        image: confluentinc/cp-zookeeper
+        restart: unless-stopped
+        container_name: zookeeper
+        ports:
+          - "2181:2181"
+        environment:
+            ZOOKEEPER_CLIENT_PORT: 2181
+            ZOOKEEPER_TICK_TIME: 2000
+        networks:
+            - kafka
+
+    broker: # broker = server running kafka instance
+        image: confluentinc/cp-kafka
+        restart: unless-stopped
+        container_name: broker
+        ports:
+            - "9092:9092"
+        depends_on:
+            - zookeeper
+        environment:
+            KAFKA_BROKER_ID: 1
+            KAFKA_ZOOKEEPER_CONNECT: "zookeeper:2181"
+            KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: PLAINTEXT:PLAINTEXT, PLAINTEXT_INTERNAL:PLAINTEXT
+            KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://localhost:9092,PLAINTEXT_INTERNAL://broker:29092
+            KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
+            KAFKA_TRANSACTION_STATE_LOG_MIN_ISR: 1
+            KAFKA_TRANSACTION_STATE_LOG_REPLICATION_FACTOR: 1
+        networks:
+            - kafka
