+++ apps/reviews-api/go.mod
@@ -2,15 +2,18 @@ module github.com/stamford-syntax-club/course-compose/reviews
 
 go 1.21.2
 
-replace github.com/stamford-syntax-club/course-compose/prisma => ../prisma
-
 require (
+	github.com/confluentinc/confluent-kafka-go v1.9.2
 	github.com/gofiber/contrib/jwt v1.0.7
 	github.com/gofiber/fiber/v2 v2.51.0
 	github.com/golang-jwt/jwt/v5 v5.0.0
+	github.com/iancoleman/strcase v0.3.0
 	github.com/joho/godotenv v1.5.1
-	github.com/stamford-syntax-club/course-compose/prisma v0.0.0-00010101000000-000000000000
+	github.com/shopspring/decimal v1.3.1
+	github.com/steebchen/prisma-client-go v0.28.0
 	github.com/stretchr/testify v1.8.4
+	github.com/takuoki/gocase v1.1.1
+	golang.org/x/text v0.14.0
 )
 
 require (
@@ -18,20 +21,15 @@ require (
 	github.com/andybalholm/brotli v1.0.5 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/google/uuid v1.4.0 // indirect
-	github.com/iancoleman/strcase v0.3.0 // indirect
 	github.com/klauspost/compress v1.16.7 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mattn/go-runewidth v0.0.15 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/rivo/uniseg v0.2.0 // indirect
-	github.com/shopspring/decimal v1.3.1 // indirect
-	github.com/steebchen/prisma-client-go v0.28.0 // indirect
-	github.com/takuoki/gocase v1.1.1 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasthttp v1.50.0 // indirect
 	github.com/valyala/tcplisten v1.0.0 // indirect
 	golang.org/x/sys v0.14.0 // indirect
-	golang.org/x/text v0.14.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
