+++ apps/reviews-api/review/data/datasource/db/prisma.go
@@ -1,0 +1,12 @@
+package db
+
+import "log"
+
+func New() *PrismaClient {
+	prisma := NewClient()
+	if err := prisma.Prisma.Connect(); err != nil {
+		log.Fatalln("Prisma connect: ", err)
+	}
+
+	return prisma
+}
