+++ go.work
@@ -2,6 +2,5 @@ go 1.21.2
 
 use (
 	./apps/reviews-api
-    ./apps/api-gateway
-	./apps/prisma
+	./apps/api-gateway
 )
